update image inheritance
