+++ mavimage/WebCamera.py
@@ -1,4 +1,4 @@
-import VideoCapture
+from SimpleCV import Camera
 """
 WebCamera class to be inherited by Camera class
 
@@ -8,3 +8,9 @@ class WebCamera:
 
     def __init__(self, width, height, gps, webcamera):
         # width, height, gps:GPS, webcamera:VideoCapture.device
+        self.width = width
+        self.height = height
+        self.gps = gps
+        webcamera = Camera()
+        self.webcamera = webcamera
+
